[Refactor] Events Feature
Refactor events feature, now is ready to testing
